% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgemm_C_blas.R
\name{dgemm_C_blas}
\alias{dgemm_C_blas}
\title{Description of the dgemm_C_blas function}
\usage{
dgemm_C_blas(matrix_a, matrix_b, repeats = 1)
}
\arguments{
\item{matrix_a}{matrix a}

\item{matrix_b}{matrix b}

\item{repeats}{determines how often the matrix multiplcation is repeated.}
}
\value{
returns the matrixproduct of a and b
}
\description{
This function takes two matrices a and b and performs a times b and returns
the result. The matrix multiplication can be repeated for benchmarking.
}
\examples{
set.seed(1)
a <- matrix(rnorm(20), nrow = 4, ncol = 5)
b <- matrix(rnorm(15), nrow = 5, ncol = 3)
c <- dgemm_C_blas(a, b)
}
